chore: move the sticky scaffolding to its own module
