+++ crates/matrix-sdk/src/sliding_sync/mod.rs
@@ -21,6 +21,7 @@ mod client;
 mod error;
 mod list;
 mod room;
+mod sticky_parameters;
 
 use std::{
     collections::{BTreeMap, BTreeSet},
@@ -82,83 +83,6 @@ pub struct SlidingSync {
 
     /// A lock to ensure that responses are handled one at a time.
     response_handling_lock: Arc<AsyncMutex<()>>,
-}
-
-mod sticky_parameters {
-    //! Sticky parameters are implemented in the context of a different mod, so
-    //! it's not possible write to the sticky parameters fields without
-    //! invalidating the sticky parameters.
-    //!
-    //! At the moment, the granularity of stickiness is the entire set of sticky
-    //! parameters; if needs be, we'll change that in the future to have
-    //! something that is per sticky field.
-    use ruma::{OwnedTransactionId, TransactionId};
-
-    pub trait StickyData {
-        type Request;
-
-        fn apply(&self, req: &mut Self::Request);
-    }
-
-    #[derive(Debug)]
-    pub struct StickyManager<D: StickyData> {
-        /// The data managed by this sticky manager.
-        data: D,
-
-        /// Was any of the parameters invalidated? If yes, reinitialize them.
-        invalidated: bool,
-
-        /// If the sticky parameters were applied to a given request, this is
-        /// the transaction id generated for that request, that must be matched
-        /// upon in the next call to `commit()`.
-        txn_id: Option<OwnedTransactionId>,
-    }
-
-    impl<D: StickyData> StickyManager<D> {
-        /// Create a new `StickyManager` for the given data.
-        ///
-        /// Always assume the initial data invalidates the request, at first.
-        pub fn new(data: D) -> Self {
-            Self { data, txn_id: None, invalidated: true }
-        }
-
-        /// Get a mutable reference to the managed data.
-        ///
-        /// Will invalidate the sticky set by default. If you don't need to modify the data, use `Self::data()`; if you're not sure you're going to modify the data, it's best to first use `Self::data()` then `Self::data_mut()` when you're sure.
-        pub fn data_mut(&mut self) -> &mut D {
-            self.invalidated = true;
-            &mut self.data
-        }
-
-        /// Returns a non-invalidating reference to the managed data.
-        pub fn data(&self) -> &D {
-            &self.data
-        }
-
-        /// May apply some the managed sticky parameters to the given request.
-        ///
-        /// After receiving the response from this sliding sync, the caller MUST
-        /// also call [`Self::maybe_commit`] with the transaction id from the server's
-        /// response.
-        pub fn maybe_apply(&mut self, req: &mut D::Request, txn_id: &TransactionId) {
-            if self.invalidated {
-                self.txn_id = Some(txn_id.to_owned());
-                self.data.apply(req);
-            }
-        }
-
-        /// May mark the managed data as not invalidated anymore, if the transaction id received from the response matches the one received from the request.
-        pub fn maybe_commit(&mut self, txn_id: &TransactionId) {
-            if self.invalidated && self.txn_id.as_deref() == Some(txn_id) {
-                self.invalidated = false;
-            }
-        }
-
-        #[cfg(test)]
-        pub fn is_invalidated(&self) -> bool {
-            self.invalidated
-        }
-    }
 }
 
 #[derive(Debug)]
@@ -963,45 +887,6 @@ mod tests {
         // this test also ensures that Tokio is not panicking when calling `add_list`.
 
         Ok(())
-    }
-
-    struct EmptyStickyData;
-
-    impl StickyData for EmptyStickyData {
-        type Request = bool;
-
-        fn apply(&self, req: &mut Self::Request) {
-            // Mark that applied has had an effect.
-            *req = true;
-        }
-    }
-
-    #[test]
-    fn test_sticky_parameters_api_non_invalidated_no_effect() {
-        let mut sticky = StickyManager::new(EmptyStickyData);
-
-        // At first, it's always invalidated.
-        assert!(sticky.is_invalidated());
-
-        let mut applied = false;
-        sticky.maybe_apply(&mut applied, "tid123".into());
-        assert!(applied);
-        assert!(sticky.is_invalidated());
-
-        // Committing with the wrong transaction id won't commit.
-        sticky.maybe_commit("tid456".into());
-        assert!(sticky.is_invalidated());
-
-        // Committing will validate it.
-        sticky.maybe_commit("tid123".into());
-        assert!(!sticky.is_invalidated());
-
-        // Applying without being invalidated won't do anything.
-        let mut applied = false;
-        sticky.maybe_apply(&mut applied, "tid123".into());
-
-        assert!(!applied);
-        assert!(!sticky.is_invalidated());
     }
 
     #[test]
